pyproject.toml
[project]
name = "sdt"
version = "0.1.0"
description = "A CLI tool to help stay on top of documentation"
authors = [{ name = "sudoscientific" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["gitpython (>=3.1.44,<4.0.0)", "typer (>=0.15.4,<0.16.0)"]

[project.scripts]
sdt = "sdt.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = ["pytest>=8.3.5"]
